change macos directory for machine sockets
if the machine name is long and/or the username is long, it is possible to exceed the 104byte limit for filenames in MacOS.  This PR shortcuts things to $HOME/.podman instead.

Fixes: #13609

[NO NEW TESTS NEEDED]
