+++ pkg/machine/qemu/options_darwin.go
@@ -1,13 +1,28 @@
 package qemu
 
 import (
-	"os"
+	"fmt"
+	"os/user"
+	"path/filepath"
 )
 
 func getRuntimeDir() (string, error) {
-	tmpDir, ok := os.LookupEnv("TMPDIR")
-	if !ok {
-		tmpDir = "/tmp"
+	// Because MacOS can only support 104byte filenames, we abandon the
+	// long directory names and simply use ~/.podman
+	systemUser, err := user.Current()
+	if err != nil {
+		return "", err
 	}
-	return tmpDir, nil
+	return systemUser.HomeDir, nil
+}
+
+func (v *MachineVM) getSocketandPid() (string, string, error) {
+	rtPath, err := getRuntimeDir()
+	if err != nil {
+		return "", "", err
+	}
+	socketDir := filepath.Join(rtPath, ".podman")
+	pidFile := filepath.Join(socketDir, fmt.Sprintf("%s.pid", v.Name))
+	qemuSocket := filepath.Join(socketDir, fmt.Sprintf("qemu_%s.pid", v.Name))
+	return qemuSocket, pidFile, nil
 }
